core: make Float compare unequal with Rational
This commit changes the behaviour of comparisons between Float and
Rational. Before a Float and a Rational will compare equal with the ==
operator e.g.:

  >>> Float(1.0) == Integer(1)
  True

After the behaviour is that a Float is never equal to a Rational e.g.:

  >>> Float(1.0) == Integer(1)
  False

The hash of Float is fixed to be consistent with that of float so e.g.

  >>> F = Float(2.3)
  >>> f = 2.3
  >>> f == F
  True
  >>> hash(f)
  691752902764107778
  >>> hash(f) == hash(F)
  True

This means that the hash for Float satisfies the hash invariant:

  A == B   -->   hash(A) == hash(B)

This is possible since in this commit a Float only compares equal to a
float and not to any other sympy object.

This commit also means that a Float will only compare equal to another
Float if they have the same precision e.g.:

  >>> Float(2, 3) == Float(2, 10)
  False

Taken together this means that the hash invariant is always satisfied
and that structural equality is satisfied whan comparing Floats with
other sympy objects. What is lost is the transitivity of == when mixing
sympy types and Python types e.g.:

  >>> Float(2.0) != Integer(2) == 2 == 2.0 == Float(2.0)

This is not a backwards compatible change. This commit is the first of
many that will be needed just to get the sympy codebase into a coherent
state.